Enable clang-tidy on header files
clang-tidy appears to search for the .clang-tidy configuration file in
the current working directory before searching the parent directories
of the source file. Previously a dummy .clang-tidy file was placed in
the build directory with the intent of avoiding running clang-tidy on
external projects like Google Test. Unfortunately, since the build
directory was the current working directory at build time this meant
that header files were not checked by clang-tidy. This change fixes the
problem by putting the dummy .clang-tidy file in the _deps directory
that CMake uses for external projects.

To ensure that headers in new directories such as kleidicv_thread are
checked by clang-tidy, make the header filter regular expression match
everything.

+++ kleidicv/src/remap/remap_sc.h
@@ -268,8 +268,9 @@ class RemapS16Point5SVE2<uint8_t> {
     acc_b = svmlalb_u32(acc_b, line1, yfrac);
     acc_t = svmlalt_u32(acc_t, line1, yfrac);
 
-    svuint16_t result = svshrnt(svshrnb(acc_b, 2 * REMAP16POINT5_FRAC_BITS),
-                                acc_t, 2 * REMAP16POINT5_FRAC_BITS);
+    svuint16_t result = svshrnt(
+        svshrnb(acc_b, static_cast<uint64_t>(2 * REMAP16POINT5_FRAC_BITS)),
+        acc_t, static_cast<uint64_t>(2 * REMAP16POINT5_FRAC_BITS));
     svst1b_u16(pg, &dst[0], result);
     mapfrac += step;
     dst += step;
